+++ src/index.js
@@ -16,6 +16,14 @@ export { locateViewer, locateViewerPrecise, timeZoneLocation, countryLocation, p
 export { recordCanvas, downloadBlob, canRecord, supportedRecordingType } from "./recorder.js";
 export { SphereTexture } from "./texture.js";
 export { Media, drawFitted } from "./media.js";
-export { DEFAULT_LICENSE_KEY, inspectLicenseKey, hasLicenseKey } from "./license.js";
+export {
+  DEFAULT_LICENSE_KEY,
+  LICENSE_KEY_PREFIX,
+  LICENSE_PAGE_URL,
+  inspectRuntime,
+  inspectLicenseKey,
+  verifyLicenseKey,
+  hasLicenseKey,
+} from "./license.js";
 export { mapAspect, colorScale, subsolarPoint, greatCircle, angularDistance, pointInGeometry, geometryBounds, projections } from "./geo.js";
 export { default } from "./geo-globe.js";